Refactor user DTOs and remove unused migration files. Added Longitude property to zFieldDTO and zField models. Updated index.html and other HTML files with new state data. Cleaned up various unused assets and scripts.

+++ Dtos/zsk/zFieldDTO.cs
@@ -9,7 +9,7 @@ namespace BackEndGasApp.Models.zsk
     {
         public int zFieldId { get; set; }
         public string Name { get; set; }
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
+        public decimal Latitude { get; set; }
+        public decimal Longitude { get; set; }
     }
 }
